#1158 Refactoring PeerReview Types
