+++ components/PeerReviews/config/PeerReviewInviteAPI.tsx
@@ -1,6 +1,6 @@
 import { Helpers } from "@quantfive/js-web-config";
 import API from "~/config/api";
-import { PeerReviewRequest, parsePeerReviewRequest } from "./PeerReviewTypes";
+import { PeerReviewRequest, parsePeerReviewRequest } from "~/config/types/peerReview";
 import { captureEvent } from "~/config/utils/events";
 
 type Args = {
