+++ components/PeerReviews/config/PeerReviewRequestAPI.ts
@@ -1,6 +1,6 @@
 import { Helpers } from "@quantfive/js-web-config";
 import API from "~/config/api";
-import { PeerReviewRequest, parsePeerReviewRequest } from "./PeerReviewTypes";
+import { PeerReviewRequest, parsePeerReviewRequest } from "~/config/types/peerReview";
 
 type Args = {
   onError: Function;
