+++ config/types/root_types.ts
@@ -2,6 +2,32 @@ export type ID = string | number | null | undefined;
 export type KeyOf<ObjectType> = keyof ObjectType;
 export type ValueOf<ObjectType> = ObjectType[keyof ObjectType];
 
+export type Post = {
+  id: ID,
+  title?: string,
+  slug?: string,
+}
+
+export type UnifiedDocument = {
+  id: ID,
+  documentType: string,
+  document?: Post,
+}
+
+export type AuthorProfile = {
+  id: ID,
+  profileImage?: string,
+  firstName?: string, 
+  lastName?: string, 
+}
+
+export type RHUser = {
+  authorProfile?: AuthorProfile,
+  id: ID,
+}
+
+// TODO: Deprecate this in favor of RHUser
+// NOTE: Avoid using this type
 export type User = {
   agreed_to_terms: boolean;
   author_profile: {
@@ -77,3 +103,56 @@ export type AuthStore = {
   uuid: null | string;
   walletLink: any; // TODO
 };
+
+export const parseUnifiedDocument = (raw: any): UnifiedDocument => {
+  if (typeof(raw) !== "object") {
+    return raw;
+  }
+
+  const parsed = {
+    id: raw.id,
+    documentType: raw.document_type,
+  }
+
+  if (Array.isArray(raw.documents)) {
+    parsed["document"] = parsePost(raw.documents[0]);
+  }
+
+  return parsed;
+}
+
+export const parsePost = (raw: any): Post => {
+  if (typeof(raw) !== "object") {
+    return raw;
+  }
+
+  return {
+    id: raw.id,
+    title: raw.title,
+    slug: raw.slug,
+  }
+}
+
+export const parseAuthorProfile = (raw: any): AuthorProfile => {
+  if (typeof(raw) !== "object") {
+    return raw;
+  }
+
+  const parsed = {
+    id: raw.id,
+    profileImage: raw.profile_image,
+    firstName: raw.first_name,
+    lastName: raw.last_name,
+  }
+
+  return parsed;
+}
+
+export const parseUser = (raw: any): RHUser => {
+  const parsed = {
+    id: raw.id,
+    authorProfile: parseAuthorProfile(raw.author_profile)
+  }
+
+  return parsed;
+}
